New Aalto Fuksis Telegram bot server
This bot now runs continuously and listens to the incoming commands.

In this version, `/canteens` command is implemented. This command
displays all the information about students restaurants located in
Otaniemi. User can freely interact with this bot using inline buttons.
In the end, the command content can be either send to the group chat, or
deleted.